chore: run fibre txsim inside the servers
