+++ Makefile
@@ -360,12 +360,13 @@ txsim-build-docker:
 	docker build -t ghcr.io/celestiaorg/txsim -f docker/txsim/Dockerfile  .
 .PHONY: txsim-build-docker
 
-## build-talis-bins: Build celestia-appd, txsim, and latency-monitor binaries for talis VMs (ubuntu 22.04 LTS)
+## build-talis-bins: Build celestia-appd, txsim, latency-monitor, and fibre-txsim binaries for talis VMs (ubuntu 22.04 LTS)
 build-talis-bins:
 	mkdir -p build
 	GOOS=linux GOARCH=amd64 go build -ldflags "$(LDFLAGS_STANDALONE)" -o build/celestia-appd ./cmd/celestia-appd
 	GOOS=linux GOARCH=amd64 go build -o build/txsim ./test/cmd/txsim
 	GOOS=linux GOARCH=amd64 go build -o build/latency-monitor ./tools/latency-monitor 2>/dev/null || echo "latency-monitor not found, skipping"
+	GOOS=linux GOARCH=amd64 go build -o build/fibre-txsim ./tools/fibre-txsim
 .PHONY: build-talis-bins
