+++ tools/talis/genesis.go
@@ -27,6 +27,7 @@ func generateCmd() *cobra.Command {
 		nodeBinaryPath                string
 		txsimBinaryPath               string
 		latencyMonitorBinaryPath      string
+		fibreTxsimBinaryPath          string
 		observabilityDirPath          string
 		useMainnetStakingDistribution bool
 	)
@@ -102,6 +103,11 @@ func generateCmd() *cobra.Command {
 				if err := copyFile(latencyMonitorBinaryPath, filepath.Join(buildDest, "latency-monitor"), 0o755); err != nil {
 					log.Printf("failed to copy latency monitor binary: %v", err)
 				}
+
+				// Copy fibre-txsim binary
+				if err := copyFile(fibreTxsimBinaryPath, filepath.Join(buildDest, "fibre-txsim"), 0o755); err != nil {
+					log.Printf("failed to copy fibre-txsim binary: %v", err)
+				}
 			}
 
 			if err := writeAWSEnv(filepath.Join(payloadDir, "vars.sh"), cfg); err != nil {
@@ -134,6 +140,7 @@ func generateCmd() *cobra.Command {
 	cmd.Flags().StringVarP(&nodeBinaryPath, "node-binary", "n", filepath.Join(gopath, "celestia"), "node binary to include in the payload (assumes the binary is installed")
 	cmd.Flags().StringVarP(&txsimBinaryPath, "txsim-binary", "t", filepath.Join(gopath, "txsim"), "txsim binary to include in the payload (assumes the binary is installed)")
 	cmd.Flags().StringVar(&latencyMonitorBinaryPath, "latency-monitor-binary", filepath.Join(gopath, "latency-monitor"), "latency monitor binary to include in the payload")
+	cmd.Flags().StringVarP(&fibreTxsimBinaryPath, "fibre-txsim-binary", "f", filepath.Join(gopath, "fibre-txsim"), "fibre-txsim binary to include in the payload")
 	cmd.Flags().StringVar(&observabilityDirPath, "observability-dir", "", "path to observability directory containing docker-compose, Prometheus config, and scripts (required if observability nodes are configured)")
 	cmd.Flags().BoolVarP(&useMainnetStakingDistribution, "mainnet-staking-distribution", "m", false, "replace the default uniform staking distribution with the actual mainnet distribution")
